添加用户认证信息，post用户认证以及外键user

+++ app/router/post.py
@@ -10,24 +10,25 @@ router = APIRouter(
 
 #  查询接口
 @router.get("/",response_model = List[schemas.Post])
-def get_post(db:Session = Depends(get_db)):
+def get_post(db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # 使用orm查询数据库表
     posts = db.query(models.Post).all()
     return posts
 
 # 创建接口
 @router.post("/",response_model=schemas.Post)
-def create_post(post :schemas.PostCreate,db:Session = Depends(get_db),user_id: int = Depends(oauth2.get_current_user)):
+def create_post(post :schemas.PostCreate,db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # 使用orm创建数据信息
-    posts = models.Post(title= post.title,content=post.content,published=post.published)
+    print(current_user)
+    posts = models.Post(title= post.title,content=post.content,published=post.published,owner_id=current_user.id)
     db.add(posts)
     db.commit()
     db.refresh(posts)
     return posts
 
 # 查看单独接口
-@router.get("/{id}",response_model=schemas.Post)
-def get_one_post(id:int,response:Response,db:Session = Depends(get_db)):
+@router.get("/{id}",response_model=schemas.Post,)
+def get_one_post(id:int,response:Response,db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # 使用orm查看单独数据信息
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
@@ -37,25 +38,31 @@ def get_one_post(id:int,response:Response,db:Session = Depends(get_db)):
 
 # 删除post中的数据
 @router.delete("/{id}",status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(id:int,response:Response,db:Session = Depends(get_db)):
+def delete_post(id:int,response:Response,db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     # 使用orm删除数据信息
     post = db.query(models.Post).filter(models.Post.id==id).first()
     if not post:
         response.status_code = status.HTTP_404_NOT_FOUND
         return {"data":"post not found"}
+    if post.owner_id != current_user.id:
+        response.status_code = status.HTTP_400_BAD_REQUEST
+        return {"data":"you are not the user of this post"}
     db.delete(post)
     return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 
 # 更新post中的数据
 @router.put("/{id}",response_model = schemas.Post)
-def put_post(id:int,post:schemas.UpdateCreate,response:Response,db:Session = Depends(get_db)):
+def put_post(id:int,post:schemas.UpdateCreate,response:Response,db:Session = Depends(get_db),current_user: int = Depends(oauth2.get_current_user)):
     #使用orm进行数据更新
     posts_query = db.query(models.Post).filter(models.Post.id == id)
     posts_db= posts_query.first()
     if not posts_db:
         response.status_code = status.HTTP_404_NOT_FOUND
         return {"data":"post not found"}
+    if posts_query.owner_id != current_user.id:
+        response.status_code = status.HTTP_400_BAD_REQUEST
+        return {"data":"you are not the user of this post"}
     posts_query.update(post.dict(), synchronize_session=False) 
     db.commit()
     db.refresh(posts_db)
